Stop using Docker when executing the build
This change modifies the Ansible bootstrap logic to install all of the
necessary build dependencies onto the host that's running Ansible. This
way, we can execute appliance-build directly on the host, rather than
relying on the Docker contianer. Additionally, the documentation has
been updated to no longer reference the use of Docker.

With that said, we still build the Docker image and support building in
the Docker contianer just as before, so we can temporarily maintain
backwards compatibility with our automation, and avoid flag days.